Entrega 19

+++ frontend/src/components/Listado.js
@@ -6,17 +6,35 @@ import {useContext} from "react"; */
 import Tarjeta from "./Tarjeta";
 const {config} = require('../config/index.js');
 const BASE_URL = config.base_url;
-function Listado() {
+function Listado(user) {
+    let carrito = user.user.carrito
+    user  = user.user.user
     const [productos, setProductos] = useState([]);
     const [loader, setLoader] = useState(true);
-   // const {carritoId} = useContext(AppContext);
-  
+    const [isAdmin, setIsAdmin] = useState(false);
+    
     useEffect(() => {
         setTimeout(() => {
-        getLista();
+            comprobarAdmin(user);
+            getLista();
         }, 1000);
-    }, []);
+    }, [user, carrito]);
  
+    const comprobarAdmin = (user) => {
+            fetch(`${BASE_URL}/auth/check/${user.email}`, {
+                method: 'GET',
+                headers: {
+                    'Content-Type': 'application/json'
+                    }
+                    })
+                    .then(res => res.json())
+                    .then(data => {
+                            setIsAdmin(data.isAdmin);
+                        }
+                    
+                )
+    }
+
     const getLista = () => {
         fetch(`${BASE_URL}/api/productos`)
         .then((response) => {
@@ -32,35 +50,45 @@ function Listado() {
         });
     };
 
-   
     const eliminarProducto = (id) => {
-        fetch(`${BASE_URL}/api/productos/${id}`, {
+        if(isAdmin){
+            fetch(`${BASE_URL}/api/productos/${id}`, {
             method: "DELETE"
         })
             .then(res => res.json())
             .then(data => {
                 console.log(data);
-                setProductos(productos.filter(productos => productos.code !== id));
+                setProductos(productos.filter(productos => productos._id !== id));
                 window.location.reload();
-            })
+            })}
+            else{
+                alert("No tienes permisos para eliminar productos");
+            }
     }
 
   
     const editarProducto = (productos) => {
         window.location.href = `/edit/${productos._id}`;
     }
-/*
-    const agregarAlCarrito = (id) => {console.log(carritoId)
-        fetch(`/api/carrito/${carritoId}/productos/${id}`, {
+
+    const agregarAlCarrito = (id) => {
+        fetch(`${BASE_URL}/api/carrito/${user._id}/productos/${id}`, 
+        {
             method: "POST"
-        }) 
+        }).then(res => res.json())
+            .then(data => {console.log( `${BASE_URL}/api/carrito/${user._id}/productos/${id}`);})
+            .catch(err => {
+                console.log(err);
+                
+            }
+            )
+    }
 
-        }*/
 
-  
+        
 
     return (
-        <>
+        <div className="container">
         {loader ? (
         <>  <button className="btn btn-light" type="button" disabled>
             <span className="spinner-grow spinner-grow-sm" role="status"  aria-hidden="true"></span>
@@ -76,14 +104,17 @@ function Listado() {
         ) : (<>
         
         {productos.length > 0 ?
-            ( <>{productos.map(producto => (
+            ( <div className="row">
+            {productos.map(producto => (
                     <Tarjeta
                         key={producto._id}
+                        isAdmin={isAdmin}
                         productos={producto}
                         eliminarProducto={eliminarProducto}
-                        editarProducto={editarProducto}/>
-                
-                ))}</>
+                        editarProducto={editarProducto}
+                        agregarAlCarrito={agregarAlCarrito}/>
+                ))}
+            </div>
             )
             : (
              <div className="col-5">
@@ -91,20 +122,14 @@ function Listado() {
                      <div className="card-body">
                          <div className="card-text">
                                     <h4>No hay productos</h4>
-                                    <button className="btn btn-primary" type="button" onClick={() => window.location.href = "/agregar"}>Agregar</button>
                             </div>
                         </div>
                     </div>
                 </div>
             )}
-
-        
          </>) }
-        </>
+        </div>
     );
 }
 
-
-
-
-export default Listado;
+export default Listado;
